Modify role.delete to no-op if role missing
If "kappa delete" fails midway then re-running it will fail during
role removal.

This PR modifies `delete` to check if the role exists.  If it does not
then we log a debug line and return early.

I also consolidated various methods that were calling `get_role` so that
error handling is consistent, and removed `_find_all_roles` as
`get_role` is sufficient, and probably faster (particularly for accounts
with many roles).