LambdaFunctionArn shouldn't be suffixed with the environment
When configuring the S3 bucket event the function `add` builds the `LambdaFunctionConfigurations` which specifies the `LambdaFunctionArn`. The ARN should be the functions *exact* ARN and not suffixed by the environment ex.`arn:aws:lambda:us-west-1:1234567890123:function:bigquery_commit:main-env`.

It makes the Lambda UI and the S3 event sources UI confused as they don't recognize that the specific Lambda function even exists.

The reason why I made this change is because I had problems applying the bucket event cross-account and could not do this without this change.